Update patch set 1
Patch Set 1:

(3 comments)

> Patch Set 1:
> 
> (1 comment)
> 
> kinda feels like too large a patch to patch into the library.  Other thoughts about what we might do?  I assume the issue is that the library tries to alloc too much memory?  or is PDFium the location of the large alloc?

I am not sure if there is really a smaller way to achieve this. This is a single function added to the API that doesn't need to modify anything else. I first attempted to do these calculations in the PDFium code, but it requires accessing a bunch of types that are hidden in the library, so I would have to make them visible.

Other thoughts I had for attempting to resolve this. Threading into the library calls a memory limit parameter, so when it goes to decode the stream it will fail if it is too big. Alternatively changing how the low level allocation works to check for a hard limit. The problem with these approaches is that they will require some bookkeeping since there is more then one allocation point.

The core issue is that the decode step in the library will allocate a large amount of memory and spend time decode the image, then we will immediate fail the entire process because we cannot allocate a bitmap. This is inefficient and is generating false-ish positives, since if we fail before the allocation step the fuzzers are happy.

Patch-set: 1
Reviewer: Ryan Harrison <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0