Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Can you explain "We immediately get killed"?  Did we start step 6?

LibFuzzer reports out of memory, since the fuzzers run with a limit of 2GB if I  recall.

It will normally be between steps 5 and 6, since step 5 caused a large allocation with in the library code. If you are running without LibFuzzer/a memory limit the stream decode will fail on step 6 since that code checks how large the bitmap it is going to allocate before doing it.

So this change will eliminate 'false' positives where the library allocates a bunch of memory, but we were never going to use it, since we couldn't write it out, but the because it is running in the fuzzer it causes an OOM. This case is likely to be pretty common, since a 120 byte stream can cause a 6 GB allocation, . A header without any tiles is considered a valid image, just filled with the default colour. In production there will be a efficiency gain from not allocating/decoding/deallocating in the stream decode fails case.

Patch-set: 1
Reviewer: Ryan Harrison <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0