Move decompressing of JPX out of Init and into Decode
This is a re-landing of
https://pdfium-review.googlesource.com/c/pdfium/+/13550, which needed
to be partially reverted due to https://crbug.com/764177. The major
difference in this CL is addition of pdfium_get_components() to the
libopenjpeg API, which determines the expected number of components
from decoding the image. This varies from what is reported in the
header, since if PCLR is applied the number of components will be the
number of channels, which may not be equal to what was specified in
the header.

From the original CL:

In the existing implementation of the JPX decoder, Init extracts the
header from the image and then immediately decodes the rest of
it. This is problematic if it is a very large image that we won't be
able to allocate a bitmap for. The code has been changed to instead
delay decoding until the Decode method, so that if a bitmap cannp
cannot be allocated decoding won't be attempted.

BUG=chromium:754423,chromium:761005

Change-Id: I6fff0a3d84e148e05399932577d3304b9c6c36e9