Cleanup FindNextBreakPos variables
The ePreType flag is only used if bFirst is true. After the first
iteration of the loop bFirst is always false. This CL removes the
updating of ePreType as it will never be used.

Also bFirst is only set false at the end of the loop and the do {} while is
changed into a while {}

Change-Id: Iaced1b38bddcc6f5483ae20993ac69c93b2e3f97