Confirm that a global palette has data before attempting to use it
Previous implementation assumed that if the local colour palette was
not specified and the global palette had its size specified, then use
the global. If the global palette is disable, it will not have data,
but it may have a size. Technically the size is giberish in this case,
but the value is allowed to be non-zero, so isn't a sufficient check.

BUG=chromium:768089

Change-Id: Iaec15fcd65f3983056df7d56d29118a516334cd9