Update patch set 4
Patch Set 4:

When is the next branch?  We may be able to make our lives easier when it comes to merging patches if we align this immediately before/after a branch is taken, though I'm not convinced which is better or if it matters.  Thoughts?

Patch-set: 4
Reviewer: Tom Sepez <5495@fca2d64a-0420-3532-b200-12848c95a47b>