Make names of GIF types less opaque
A lot of the structs in the GIF code use acronyms that are non-obvious
unless you are familar with the GIF spec, and even then lead to lots
of looking back. Making the naming more explicit. Also converted
GifImage to be a struct, since it didn't have any methods defined for
it.

BUG=pdfium:903

Change-Id: I97ca2adfdc4ada8bcaef5f8f317c1dea7365d70d