Replace GetWord_LSBFirst with FXWORD_GET_LSBFIRST
The existing implementation of a LSB first word method was incorrect,
in addition it was implemented to the BMP code, but also used in the
GIF code. Thus is should be moved to a common location.

Also added in an implementation for FXWORD_GET_MSBFIRST, since the
GIF code will need this.

BUG=pdfium:914

Change-Id: I0e84813356fbd456b293a190da3c2cde01a6580b