Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Commit-Queue+1
> > > 
> > > Thoughts? Best idea or Worlds Best idea?
> > > 
> > > Was figuring I'd do this by directory as it makes the reviews a lot easier.
> > 
> > Structs don't have the trailing underscore. :-P
> > 
> > If this is mostly automated, then sure, why not. Blink did it and they probably have even more code. Probably need to patch up things manually afterwards. e.g. d_word_foo -> dword_foo.
> 
> 
> Chatting with tsepez@ their preference is to do this all in one big CL instead of a pile of smaller CLs. Which do other people prefer?
> 
> If it's one big CL, will probably mean getting the Blink tools working as they're smarter then my script, heh.

If you can generate it and get it submitted before anything else lands and causes a conflict, that would be my preference. however, if we don't have a reliable too to generate the giant CL without any adjustments, it would be worth doing it in medium pieces - like a directory per CL - so that smaller and manageable conflicts happens but CLs are still fast to generate.

Patch-set: 1
Reviewer: Henrique Nakashima <7035@fca2d64a-0420-3532-b200-12848c95a47b>