Make GIF decoder more standards complient
Fixed issue with unit tests that was causing raw data to be backwards
and reverted related LSB -> MSB change that was introduced due to
this.

If global palette not set then the background colour index should be
0.

Check that background colour index is valid when global palette
exists.

Check if transparency index is valid for the palette of the frame it
is being applied to.

BUG=chromium:770337

Change-Id: I5d9b648f45bbb4c93218664a7035e4d01dbeb627