Convert XFA Doc Types to be more precise
The existing types are PDF, Dynamic & Static, which are poorly named
since they don't really convey the fundamental differences between the
types. Instead PDF is confusing because PDFium only handles PDFs, and
Dynamic & Static describe what a form may do, not how it is specified
or some other fundamental difference. The terms Dynamic and Static
were especially confusing, since XFAF documents must be static,
i.e. not change their layout, whereas Full documents could be either
Dynamic or Static, since they could change their layout, but were not
required to.

The types have been renamed to be clear that they are talking about
the type of PDF document being described. PDF becomes None, since this
is used to indicate that there are no XFA forms in the
document. Dynamic becomes Full, since this indicates that the entire
XFA spec is being used for the forms, specifically display layout is
in the XML. Static has become Foreground, since the form is specified
using the XFAF (XFA Foreground) subset of the spec.

The terms Full & Foreground come from the XFA spec. I would have
prefered XFAF to have a different name, since it is the
display/foreground layer that isn't XFA in it.

BUG=pdfium:917

Change-Id: I4335958c4a11d77d3bbe63b93602dd5bc14acb57