Convert XFA Doc Types to be more precise
The existing types are PDF, Dynamic & Static, which are poorly named
since they don't really convey the fundamental differences between the
types. "PDF" is confusing because PDFium only handles PDFs, and
"Dynamic" & "Static" describe what a form may do, not how it is 
specified or some other fundamental difference. The terms "Dynamic"
and "Static" were especially confusing, since XFAF documents must be 
static by definition, whereas full XFA documents can be either
dynamic or static, depending on whether or not they change their 
layout.

The types have been renamed to be clear that they are talking about
the type of PDF document being described. "PDF" becomes "None", since 
this is used to indicate that there are no XFA forms in the
document. "Dynamic" becomes "Full", since this indicates that the 
entire XFA spec is being used for the forms, specifically display 
layout is in the XML. "Static" has become "Foreground", since the form 
is specified using the XFAF (XFA Foreground) subset of the spec.

The terms Full & Foreground come from the XFA spec. I would have
preferred XFAF to have a different name, since it is the
display/foreground layer that isn't XFA when using it.

BUG=pdfium:917

Change-Id: I4335958c4a11d77d3bbe63b93602dd5bc14acb57