Remove ContrastAdjust().
This post-processing increased the contrast of scaled-down features -
mainly lines and fonts, relative to the background. The effect does not
improved readability and in some cases makes the scaled-down version
look like a different document at a glance. Text that is normal
weight appears bold when scaled down in these cases.

Change-Id: I2544d45e2bcec25d6742d2a60c1316d8df08cce1