Improve failed decoding in CJBig2_ArithDecoder
This CL adds a bool |m_DecodeError| to CJBig2_ArithDecoder to keep
track of whether the decoding has failed or not. It also replaces the
method IsComplete() with ShouldNoLongerDecode() using this new
attribute. Note that DECODE() can attempt to read past the current
value being decoded, so a decoding error should only prevent the next
DECODE from happening (the current DECODE should be accepted).

Bug: 767156
Change-Id: I335d3898ce1a46316025b5179b62d47a59ebeaee