Consistently apply JS size check
This CL makes the use of CXFA_IsTooBig consistent and universal across
all of the ToJavascript and related methods. Previously this method
was only applied in some calls.

It is no being tested as a precondition on every call and as a post
condition for the entire translation process.

It has been removed as a post condition for individual emission calls,
instead letting the pre-conditions optimize for early exit and have
the final post condition check for correctness.

This resolves the specific crash in the bug listed, but there is still
issues in executing the JS that needs to be investigated.

BUG=chromium:752920

Change-Id: I1a8bfe21e0a17c2e47592fc6017060243674f1bc