Reland of "Remove ContrastAdjust()."
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/742061
with a minor non-functional difference in an embedder test.

The cl was reverted because it created differences with gold files
in a chromium test, blocking the pdfium deps roll.

The new gold files will be submitted in
https://chromium-review.googlesource.com/c/chromium/src/+/742061
which will also be amended with the hash of this commit to roll it
manually immediately.
Change-Id: If6aa6a7329f2bb6ecb2331f3e850988e3af27a49