Update patch set 3
Patch Set 3:

> Patch Set 3:
> Filed https://crbug.com/pdfium/926 to track embedder work.

Thanks. When I get to my home computers, I can check the data dirs for the native PDF viewers to see how much persistence data exists, as a sample of how much real world use this feature gets.

Another thing we want to persist, which I believe we already have a bug for, is what page to jump to when restoring a tab with a PDF in it.

Patch-set: 3