Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Any chance we can reduce linearized.pdf in size a little? It's mostly embedded fonts.
> 
> i.e. Make it more like rectangles.pdf.
> 
> My other suggestion is to render both linearized.pdf and the saved copy, and make sure CompareBitMap() works with the same hash for both.

Replaced the text with some circle annotations to avoid embedding fonts. This reduces the file size to 1/3.

I'll look into rendering the saved copy, but it's not trivial because we're not actually saving the new pdf to a file yet, just writing it to memory.

Patch-set: 2
Reviewer: Henrique Nakashima <7035@fca2d64a-0420-3532-b200-12848c95a47b>