Update patch set 15
Patch Set 15:

Also, I'm fairly confident the non-test changes here are fine., but we are stuck on writing tests. So if you want to land the FPDFPage_TransFormWithClip() change first, and work on the tests later in a separate CL, that works for me.

Patch-set: 15
Reviewer: Lei Zhang <5045@fca2d64a-0420-3532-b200-12848c95a47b>