Add regression tests for issues with correctly removing hyphens
There was a regression due to a refactor, where the public API was no
longer removing soft hyphens for line broken words. This was causing
issues with find and copy/paste operations that depend on selecting a
region of text. This change is covered by
FPDFTextEmbeddertest.GetTextWithHyphen.

FPDFTextEmbeddertest.bug_782596 is a regression test for a bug that was
introduced by the original fix. It only fails when running the test under
ASAN.

BUG=pdfium:935

Change-Id: I26096583c35f9246a3662e702f89b742f1146780