Convert from character stream to visible text indices in GetText
Most of the API methods FPDFText operate on indices in terms of the
underlying stream of characters. This stream includes non-printing
control characters, which are not part of the visible text. The
majority of files do not appear to have these hidden characters so
there is a 1:1 correspondence between them. When they are present
conversion needs to occur to make sure that GetText doesn't attempt to
retrieve for a span that is out of range.

BUG=chromium:788103,chromium:788220

Change-Id: I4c9fa403ea65754ba94e3f15ded49fe0641e9db5