Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> (2 comments)
> 
> Should we wait for the red bots to get resolved? Should we check binary size increase from this CL?


I'm working on the 3 failing tests, not sure what's going wrong at the moment. I can check the binary size change, but there are going to be a few of these CLs as I move stuff from the giant arrays into the classes and I don't think we want to stop in the middle. So, I guess we either continue on, or I try to get everything converted in a branch and get the final size?

Patch-set: 2
Reviewer: dsinclair <5965@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0