Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> tests?

I agree this should be tested, but I am honestly not sure if we have testing infrastructure for this. I can verify in Chrome that the text is correct now.

Specifically this text appears in a message box/pop up that is created when validation has failed. Do we have a mechanism for testing this content?

Patch-set: 1
Reviewer: Ryan Harrison <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0