Speed up test for bug 765384.
It has a gc() JS function that performs a loop with many iterations.
Under MSAN, V8 runs in a simulator which is slow and requires a long
time to simulate this loop. So speed this up this test by reducing the
number of iterations in gc(). The original bug still repros consistently
with the modified test when the bug fix gets reverted.

BUG=chromium:765384,pdfium:960

Change-Id: If8aa21dfa5a78239bdc44d26c8348739c1e005de