Speed up test for bug 765384.
It has a gc() JS function that performs a loop with many iterations.
Under MSAN, V8 runs in a simulator which is slow and requires a long
time to simulate this loop. Work around this by adding a variant of the
test PDF. The "fast" test PDF calls V8's built in gc(), which works
since EmbedderTest exposes gc(). The original "slow" test PDF remains
for use by others.

BUG=chromium:765384,pdfium:960

Change-Id: If8aa21dfa5a78239bdc44d26c8348739c1e005de