Open FPDFDest_GetView API.
FPDFDest_GetView returns the view fit type and the parameters of the
view for a given destination.

This is useful to have more precise internal links and bookmarks that
are able to manipulate the viewport position and zoom level to focus
on the part of the PDF that it links to.

Bug: 55776, 535978, 748852
Change-Id: Ibf7df40a852030d75ec78cec7662380319569850