Update patch set 4
Patch Set 4:

In the future, if you find PDFium's toolchain to be behind Chromium's, just ask us to update it. We'll do it eventually anyway. e.g. with https://pdfium-review.googlesource.com/21370 for binutils. Sometimes the updates are straight forward, and sometimes they are a pain. Ask us to do it to spare you the pain. :)

Patch-set: 4
CC: Lei Zhang <5045@fca2d64a-0420-3532-b200-12848c95a47b>